not functionning edit admin

+++ mvc/Controllers/api.php
@@ -34,30 +34,48 @@
         }
 
     } else {    
+
         $mysqli = $db->open();
+
         if ($method == 'post'){                
-            if ($params_arr){                  
-                $mysqli->query("INSERT INTO users (name, last, email) VALUES ('$params_arr[0]','$params_arr[1]', '$params_arr[2]');");
-            }            
-        } 
-        else if ($method == 'update'){       
-            if ($params_arr){
-                
-                //$user = load_user_data($params_arr[0])->fetch;                
-                $query_result = $mysqli->query("UPDATE users SET name='$params_arr[1]' last='$params_arr[2]' WHERE id=$params_arr[0];");                    
-            }            
+            try {
+                if (count($params_arr) > 2){
+                    $user = $db->load_user_data($params_arr[2]);
+                    if ($user == null){                                  
+                        $mysqli->query("INSERT INTO users (name, last, email) VALUES ('$params_arr[0]','$params_arr[1]', '$params_arr[2]');");
+                        if (!$mysqli->error) echo "User $params_arr[2] was added successfully";
+                    } else echo "This email is already in use";
+                } else echo "Not enough params";
+            } catch (Exception $e) {
+                echo "wrong or not enough params";
+            }
+        }         
+        else if ($method == 'update' && $params_arr){                               
+            $user = $db->load_user_data_id($params_arr[0]);                                
+            if ($user) {
+                $mysqli->query("UPDATE users SET name='$params_arr[1]' last='$params_arr[2]' WHERE id=$params_arr[0];");
+                echo "User $user[email] was updated successfully";
+            }
+            else echo "Wrong ID parameter";
         }
-        else if ($method == 'delete'){        
-            if ($params_arr){
-                $db->delete_user($params_arr[0]);
-            }        
-        } else {
-            echo "Incorrect API request";
+        else if ($method == 'delete' && $params_arr){        
+            $user = $db->load_user_data_id($params_arr[0]);              
+            if ($user) {
+                $db->delete_user($params_arr[0]);                    
+                echo "User $user[email] was deleted successfully";
+            }
+            else echo "Wrong ID parameter";
+        } else echo "Incorrect API request or missing/wrong params";
+
+        if ($mysqli->error){
+            echo "Incorrect API request params";
         }
+
         $mysqli->close();
         header("Content-Type: application/json");            
         
     }
+
     
     
 ?>
